requirements.txt
asgiref==3.3.1
certifi==2020.12.5
click==7.1.2
Django==3.1.4
gunicorn==20.0.4
h11==0.11.0
image==1.5.33
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==8.0.1
python-multipart==0.0.5
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
starlette==0.14.1
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
uvicorn==0.13.1
Werkzeug==1.0.1
